Split up daemon function
