+++ pkg/daemon/daemon.go
@@ -25,20 +25,18 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
-	"net"
 	"os"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/clausecker/nfc/v2"
-	"github.com/fsnotify/fsnotify"
 	"github.com/rs/zerolog/log"
 	"github.com/wizzomafizzo/mrext/pkg/input"
 	"github.com/wizzomafizzo/tapto/pkg/config"
 	"github.com/wizzomafizzo/tapto/pkg/launcher"
 	"github.com/wizzomafizzo/tapto/pkg/platforms/mister"
 	"github.com/wizzomafizzo/tapto/pkg/tokens"
+	"github.com/wizzomafizzo/tapto/pkg/utils"
 )
 
 const (
@@ -49,108 +47,13 @@ const (
 	periodBetweenLoop  = 300 * time.Millisecond
 )
 
-type Card struct {
-	CardType string
-	UID      string
-	Text     string
-	ScanTime time.Time
-}
-
-type ServiceState struct {
-	mu              sync.Mutex
-	activeCard      Card
-	lastScanned     Card
-	stopService     bool
-	disableLauncher bool
-	dbLoadTime      time.Time
-	uidMap          map[string]string
-	textMap         map[string]string
-}
-
-func (s *ServiceState) SetActiveCard(card Card) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.activeCard = card
-	if s.activeCard.UID != "" {
-		s.lastScanned = card
-	}
-}
-
-func (s *ServiceState) GetActiveCard() Card {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.activeCard
-}
-
-func (s *ServiceState) GetLastScanned() Card {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.lastScanned
-}
-
-func (s *ServiceState) StopService() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.stopService = true
-}
-
-func (s *ServiceState) ShouldStopService() bool {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.stopService
-}
-
-func (s *ServiceState) DisableLauncher() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.disableLauncher = true
-	if _, err := os.Create(mister.DisableLaunchFile); err != nil {
-		log.Error().Msgf("cannot create disable launch file: %s", err)
-	}
-}
-
-func (s *ServiceState) EnableLauncher() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.disableLauncher = false
-	if err := os.Remove(mister.DisableLaunchFile); err != nil {
-		log.Error().Msgf("cannot remove disable launch file: %s", err)
-	}
-}
-
-func (s *ServiceState) IsLauncherDisabled() bool {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.disableLauncher
-}
-
-func (s *ServiceState) GetDB() (map[string]string, map[string]string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.uidMap, s.textMap
-}
-
-func (s *ServiceState) GetDBLoadTime() time.Time {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.dbLoadTime
-}
-
-func (s *ServiceState) SetDB(uidMap map[string]string, textMap map[string]string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.dbLoadTime = time.Now()
-	s.uidMap = uidMap
-	s.textMap = textMap
-}
-
 func pollDevice(
 	cfg *config.UserConfig,
 	pnd *nfc.Device,
-	activeCard Card,
+	activeCard Token,
 	ttp int,
 	pbp time.Duration,
-) (Card, error) {
+) (Token, error) {
 	count, target, err := pnd.InitiatorPollTarget(tokens.SupportedCardTypes, ttp, pbp)
 	if err != nil && !errors.Is(err, nfc.Error(nfc.ETIMEOUT)) {
 		return activeCard, err
@@ -159,7 +62,7 @@ func pollDevice(
 	if count <= 0 {
 		if activeCard.UID != "" && time.Since(activeCard.ScanTime) > timeToForgetCard {
 			log.Info().Msg("card removed")
-			activeCard = Card{}
+			activeCard = Token{}
 
 			if cfg.TapTo.ExitGame {
 				mister.ExitGame()
@@ -209,8 +112,8 @@ func pollDevice(
 		log.Info().Msgf("decoded text NDEF: %s", tagText)
 	}
 
-	card := Card{
-		CardType: cardType,
+	card := Token{
+		Type:     cardType,
 		UID:      cardUid,
 		Text:     tagText,
 		ScanTime: time.Now(),
@@ -219,31 +122,9 @@ func pollDevice(
 	return card, nil
 }
 
-func getSerialDeviceList() ([]string, error) {
-	path := "/dev/serial/by-id/"
-	f, err := os.Open(path)
-	if err != nil {
-		return nil, err
-	}
-	files, err := f.Readdir(0)
-	if err != nil {
-		return nil, err
-	}
-
-	var devices []string
-
-	for _, v := range files {
-		if !v.IsDir() {
-			devices = append(devices, path+v.Name())
-		}
-	}
-
-	return devices, nil
-}
-
 func detectConnectionString() string {
 	log.Info().Msg("probing for serial devices")
-	devices, _ := getSerialDeviceList()
+	devices, _ := utils.GetLinuxSerialDeviceList()
 
 	for _, device := range devices {
 		connectionString := "pn532_uart:" + device
@@ -282,7 +163,7 @@ func OpenDeviceWithRetries(config config.TapToConfig) (nfc.Device, error) {
 	}
 }
 
-func writeScanResult(card Card) error {
+func writeScanResult(card Token) error {
 	f, err := os.Create(mister.TokenReadFile)
 	if err != nil {
 		return fmt.Errorf("unable to create scan result file %s: %s", mister.TokenReadFile, err)
@@ -299,7 +180,7 @@ func writeScanResult(card Card) error {
 	return nil
 }
 
-func launchCard(cfg *config.UserConfig, state *ServiceState, kbd input.Keyboard) error {
+func launchCard(cfg *config.UserConfig, state *State, kbd input.Keyboard) error {
 	card := state.GetActiveCard()
 	uidMap, textMap := state.GetDB()
 
@@ -335,9 +216,10 @@ func launchCard(cfg *config.UserConfig, state *ServiceState, kbd input.Keyboard)
 	return nil
 }
 
-func StartService(cfg *config.UserConfig) (func() error, error) {
-	state := &ServiceState{}
+func StartDaemon(cfg *config.UserConfig) (func() error, error) {
+	state := &State{}
 
+	// TODO: this is platform specific
 	kbd, err := input.NewKeyboard()
 	if err != nil {
 		log.Error().Msgf("failed to initialize keyboard: %s", err)
@@ -351,75 +233,12 @@ func StartService(cfg *config.UserConfig) (func() error, error) {
 		state.SetDB(uids, texts)
 	}
 
-	var closeDbWatcher func() error
-	dbWatcher, err := fsnotify.NewWatcher()
+	closeDbWatcher, err := launcher.StartMappingsWatcher(
+		state.GetDBLoadTime,
+		state.SetDB,
+	)
 	if err != nil {
-		log.Error().Msgf("error creating mappings watcher: %s", err)
-	} else {
-		closeDbWatcher = dbWatcher.Close
-	}
-
-	go func() {
-		// this turned out to be not trivial to say the least, mostly due to
-		// the fact the fsnotify library does not implement the IN_CLOSE_WRITE
-		// inotify event, which signals the file has finished being written
-		// see: https://github.com/fsnotify/fsnotify/issues/372
-		//
-		// during a standard write operation, a file may emit multiple write
-		// events, including when the file could be half-written
-		//
-		// it's also the case that editors may delete the file and create a new
-		// one, which kills the active watcher
-		const delay = 1 * time.Second
-		for {
-			select {
-			case event, ok := <-dbWatcher.Events:
-				if !ok {
-					return
-				}
-				if event.Has(fsnotify.Write) {
-					// usually receives multiple write events, just act on the first
-					if time.Since(state.GetDBLoadTime()) < delay {
-						continue
-					}
-					time.Sleep(delay)
-					log.Info().Msg("database changed, reloading")
-					uids, texts, err := launcher.LoadDatabase()
-					if err != nil {
-						log.Error().Msgf("error loading database: %s", err)
-					} else {
-						state.SetDB(uids, texts)
-					}
-				} else if event.Has(fsnotify.Remove) {
-					// editors may also delete the file on write
-					time.Sleep(delay)
-					_, err := os.Stat(mister.MappingsFile)
-					if err == nil {
-						err = dbWatcher.Add(mister.MappingsFile)
-						if err != nil {
-							log.Error().Msgf("error watching database: %s", err)
-						}
-						log.Info().Msg("database changed, reloading")
-						uids, texts, err := launcher.LoadDatabase()
-						if err != nil {
-							log.Error().Msgf("error loading database: %s", err)
-						} else {
-							state.SetDB(uids, texts)
-						}
-					}
-				}
-			case err, ok := <-dbWatcher.Errors:
-				if !ok {
-					return
-				}
-				log.Error().Msgf("watcher error: %s", err)
-			}
-		}
-	}()
-
-	err = dbWatcher.Add(mister.MappingsFile)
-	if err != nil {
-		log.Error().Msgf("error watching database: %s", err)
+		log.Error().Msgf("error starting database watcher: %s", err)
 	}
 
 	if _, err := os.Stat(mister.DisableLaunchFile); err == nil {
@@ -519,81 +338,11 @@ func StartService(cfg *config.UserConfig) (func() error, error) {
 		}
 	}()
 
-	socket, err := net.Listen("unix", mister.SocketFile)
+	socket, err := StartSocketServer(state)
 	if err != nil {
-		log.Error().Msgf("error creating socket: %s", err)
+		log.Error().Msgf("error starting socket server: %s", err)
 		return nil, err
 	}
-
-	go func() {
-		for {
-			if state.ShouldStopService() {
-				break
-			}
-
-			conn, err := socket.Accept()
-			if err != nil {
-				log.Error().Msgf("error accepting connection: %s", err)
-				return
-			}
-
-			go func(conn net.Conn) {
-				log.Debug().Msg("new socket connection")
-
-				defer func(conn net.Conn) {
-					err := conn.Close()
-					if err != nil {
-						log.Warn().Msgf("error closing connection: %s", err)
-					}
-				}(conn)
-
-				buf := make([]byte, 4096)
-
-				n, err := conn.Read(buf)
-				if err != nil {
-					log.Error().Msgf("error reading from connection: %s", err)
-					return
-				}
-
-				if n == 0 {
-					return
-				}
-				log.Debug().Msgf("received %d bytes", n)
-
-				payload := ""
-
-				switch strings.TrimSpace(string(buf[:n])) {
-				case "status":
-					lastScanned := state.GetLastScanned()
-					if lastScanned.UID != "" {
-						payload = fmt.Sprintf(
-							"%d,%s,%t,%s",
-							lastScanned.ScanTime.Unix(),
-							lastScanned.UID,
-							!state.IsLauncherDisabled(),
-							lastScanned.Text,
-						)
-					} else {
-						payload = fmt.Sprintf("0,,%t,", !state.IsLauncherDisabled())
-					}
-				case "disable":
-					state.DisableLauncher()
-					log.Info().Msg("launcher disabled")
-				case "enable":
-					state.EnableLauncher()
-					log.Info().Msg("launcher enabled")
-				default:
-					log.Warn().Msgf("unknown command: %s", strings.TrimRight(string(buf[:n]), "\n"))
-				}
-
-				_, err = conn.Write([]byte(payload))
-				if err != nil {
-					log.Error().Msgf("error writing to socket: %s", err)
-					return
-				}
-			}(conn)
-		}
-	}()
 
 	return func() error {
 		err := socket.Close()
