--- pkg/launcher/launcher.go
@@ -1,88 +1,0 @@
-/*
-TapTo
-Copyright (C) 2023, 2024 Callan Barrett
-
-This file is part of TapTo.
-
-TapTo is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-TapTo is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with TapTo.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-package launcher
-
-import (
-	_ "embed"
-	"errors"
-	"io"
-	"os"
-	"strings"
-
-	"github.com/gocarina/gocsv"
-	"github.com/rs/zerolog/log"
-	"github.com/wizzomafizzo/tapto/pkg/platforms/mister"
-)
-
-type NfcMappingEntry struct {
-	MatchUID  string `csv:"match_uid"`
-	MatchText string `csv:"match_text"`
-	Text      string `csv:"text"`
-}
-
-func LoadDatabase() (map[string]string, map[string]string, error) {
-	uids := make(map[string]string)
-	texts := make(map[string]string)
-
-	if _, err := os.Stat(mister.MappingsFile); errors.Is(err, os.ErrNotExist) {
-		log.Info().Msg("no database file found, skipping")
-		return nil, nil, nil
-	}
-
-	f, err := os.Open(mister.MappingsFile)
-	if err != nil {
-		return nil, nil, err
-	}
-	defer func(c io.Closer) {
-		_ = c.Close()
-	}(f)
-
-	entries := make([]NfcMappingEntry, 0)
-	err = gocsv.Unmarshal(f, &entries)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	count := 0
-	for i, entry := range entries {
-		if entry.MatchUID == "" && entry.MatchText == "" {
-			log.Warn().Msgf("entry %d has no UID or text, skipping", i+1)
-			continue
-		}
-
-		if entry.MatchUID != "" {
-			uid := strings.TrimSpace(entry.MatchUID)
-			uid = strings.ToLower(uid)
-			uid = strings.ReplaceAll(uid, ":", "")
-			uids[uid] = strings.TrimSpace(entry.Text)
-		}
-
-		if entry.MatchText != "" {
-			text := strings.TrimSpace(entry.MatchText)
-			texts[text] = strings.TrimSpace(entry.Text)
-		}
-
-		count++
-	}
-	log.Info().Msgf("loaded %d entries from database", count)
-
-	return uids, texts, nil
-}
